+++ core-service/core-api/src/main/java/ru/home/aglar/market/core/api/OrderDto.java
@@ -2,6 +2,7 @@ package ru.home.aglar.market.core.api;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Schema(description = "Model of order")
@@ -13,7 +14,7 @@ public class OrderDto {
     @Schema(description = "List of order's items", requiredMode = Schema.RequiredMode.REQUIRED)
     private List<OrderItemDto> orderItems;
     @Schema(description = "Total price of hole order", requiredMode = Schema.RequiredMode.REQUIRED)
-    private Integer totalPrice;
+    private BigDecimal totalPrice;
     @Schema(description = "User's address", requiredMode = Schema.RequiredMode.REQUIRED)
     private String address;
     @Schema(description = "User's phone", requiredMode = Schema.RequiredMode.REQUIRED)
@@ -22,7 +23,7 @@ public class OrderDto {
     public OrderDto() {
     }
 
-    public OrderDto(Long id, String username, List<OrderItemDto> orderItems, Integer totalPrice, String address, String phone) {
+    public OrderDto(Long id, String username, List<OrderItemDto> orderItems, BigDecimal totalPrice, String address, String phone) {
         this.id = id;
         this.username = username;
         this.orderItems = orderItems;
@@ -55,11 +56,11 @@ public class OrderDto {
         this.orderItems = orderItems;
     }
 
-    public Integer getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(Integer totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
