+++ core-service/core-api/src/main/java/ru/home/aglar/market/core/api/ProductDto.java
@@ -2,6 +2,8 @@ package ru.home.aglar.market.core.api;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
+
 @Schema(description = "Model of product")
 public class ProductDto {
     @Schema(description = "Id", requiredMode = Schema.RequiredMode.REQUIRED, example = "1")
@@ -9,12 +11,12 @@ public class ProductDto {
     @Schema(description = "Title", requiredMode = Schema.RequiredMode.REQUIRED, maxLength = 255, minLength = 3, example = "Apple")
     private String title;
     @Schema(description = "price", requiredMode = Schema.RequiredMode.REQUIRED, example = "120.00")
-    private Integer price;
+    private BigDecimal price;
 
     public ProductDto() {
     }
 
-    public ProductDto(Long id, String title, Integer price) {
+    public ProductDto(Long id, String title, BigDecimal price) {
         this.id = id;
         this.title = title;
         this.price = price;
@@ -36,11 +38,11 @@ public class ProductDto {
         this.title = title;
     }
 
-    public Integer getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 }
